requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-address==0.2.5
django-allauth==0.44.0
django-autoslug==1.9.8
django-celery-beat==2.2.1
django-cleanup==5.2.0
django-colorfield==0.4.2
django-cors-headers==3.8.0
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
drf-extra-fields==3.1.1
idna==3.2
oauthlib==3.1.1
phonenumbers==8.12.31
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
